+++ podcasts/End of Year/End of Year 2025/EndOfYear2025StoriesModel.swift
@@ -1,5 +1,6 @@
 import PocketCastsDataModel
 import PocketCastsServer
+import PocketCastsUtils
 import SwiftUI
 
 class EndOfYear2025StoriesModel: StoryModel {
@@ -84,10 +85,13 @@ class EndOfYear2025StoriesModel: StoryModel {
         self.stories = stories
     }
 
+    private var firstTime = true
+
     func story(for storyNumber: Int) -> any StoryView {
+        defer { firstTime = false }
         switch stories[storyNumber] {
             case .intro:
-                return IntroStory2025()
+                return IntroStory2025(afterLoading: FeatureFlag.endOfYearLoadIsFirstStory.enabled && firstTime)
             case .numberOfPodcastsAndEpisodesListened:
                 return NumberListened2025(listenedNumbers: data.listenedNumbers, podcasts: data.top8Podcasts)
             case .topSpot:
@@ -164,7 +168,7 @@ class EndOfYear2025StoriesModel: StoryModel {
                     .resizable()
                     .frame(width: 153, height: 36)
                     .padding(.top, 16)
-                    .padding(.bottom, 26)
+                    .padding(.bottom, StoryLogoView.Constants.paddingBottom)
             })
     }