+++ podcasts/End of Year/Stories/2025/IntroStory2025.swift
@@ -1,5 +1,6 @@
 import SwiftUI
 import Lottie
+import PocketCastsServer
 
 extension Color {
     static let endOfYear2025Background = Color(hex: "28486A")
@@ -16,6 +17,17 @@ struct IntroStory2025: StoryView {
     @State private var animationProgress: AnimationProgressTime = .zero
 
     @State private var openCircle: Bool = false
+    @State private var animationFinished = false
+    @EnvironmentObject var syncProgressModel: SyncYearListeningProgress
+
+    private let afterLoading: Bool
+
+    let loadCallback: (() -> ())?
+
+    init(afterLoading: Bool, loadCallback: (() -> ())? = nil) {
+        self.loadCallback = loadCallback
+        self.afterLoading = afterLoading
+    }
 
     enum KeyFrames {
         static var circleOpen = CGFloat(110)
@@ -27,32 +39,59 @@ struct IntroStory2025: StoryView {
                 .multilineTextAlignment(.center)
                 .font(size: 25, style: .title, weight: .semibold)
                 .foregroundStyle(backgroundTextColor)
+                .padding(.horizontal, 24)
                 .mask(
                     Circle().scale(scale)
                 )
-                .opacity(opacity)
+                .opacity(animationFinished ? 0 : opacity)
         }
         .onChange(of: animationProgress) { position in
-            if position > KeyFrames.circleOpen, !openCircle {
-                openCircle = true
-                withAnimation(.easeInOut(duration: 0.005)) {
-                    self.opacity = 1
-                }
-                withAnimation(.easeInOut(duration: 1)) {
-                    self.scale = 10
+            if afterLoading {
+                self.opacity = 1
+                self.scale = 10
+            } else {
+                if position > KeyFrames.circleOpen, !openCircle {
+                    openCircle = true
+                    withAnimation(.easeInOut(duration: 0.005)) {
+                        self.opacity = 1
+                    }
+                    withAnimation(.easeInOut(duration: 1)) {
+                        self.scale = 10
+                    }
                 }
             }
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity)
         .background(
-            LottieView(animation: .named("end_of_year_2025_intro"))
-                .configure({ animationView in
-                    animationView.contentMode = .scaleAspectFill
-                })
-                .playbackMode(.playing(.fromProgress(0, toProgress: 1, loopMode: .playOnce)))
-                .getRealtimeAnimationFrame($animationProgress)
-                .scaledToFill()
-                .ignoresSafeArea()
+            ZStack {
+                if animationFinished {
+                    VStack(spacing: 15) {
+                        Spacer()
+                        ZStack {
+                            CircularProgressView(value: 1, stroke: .white.opacity(0.33), strokeWidth: 6)
+                                .frame(width: 40, height: 40)
+                            CircularProgressView(value: self.syncProgressModel.progress, stroke: .white, strokeWidth: 6)
+                                .frame(width: 40, height: 40)
+                        }
+                        Spacer()
+                    }
+                } else {
+                    LottieView(animation: .named("end_of_year_2025_intro"))
+                        .animationDidFinish({ completed in
+                            DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
+                                animationFinished = true
+                                loadCallback?()
+                            }
+                        })
+                        .configure({ animationView in
+                            animationView.contentMode = .scaleAspectFill
+                        })
+                        .playbackMode(afterLoading ? .paused(at: .progress(1)) : .playing(.fromProgress(0, toProgress: 1, loopMode: .playOnce)))
+                        .getRealtimeAnimationFrame($animationProgress)
+                        .scaledToFill()
+                        .ignoresSafeArea()
+                }
+            }
         )
         .ignoresSafeArea()
         .background(backgroundColor)
@@ -65,6 +104,6 @@ struct IntroStory2025: StoryView {
 
 struct IntroStory2025_Previews: PreviewProvider {
     static var previews: some View {
-        IntroStory2025()
+        IntroStory2025(afterLoading: false)
     }
 }