+++ podcasts/End of Year/Stories/2025/ListeningTime2025Story.swift
@@ -25,8 +25,8 @@ struct ListeningTime2025Story: ShareableStory {
 
     var formattedMinutes: String {
         let formatter = NumberFormatter()
+        formatter.locale = Locale.current
         formatter.numberStyle = .decimal
-        formatter.groupingSeparator = ","
         return formatter.string(for: Int(listeningTime / 60.0)) ?? ""
     }
 
@@ -89,7 +89,7 @@ struct ListeningTime2025Story: ShareableStory {
     func sharingAssets() -> [Any] {
         [
             StoryShareableProvider.new(AnyView(self)),
-            StoryShareableText(L10n.eoyStoryListenedToShareText(formattedMinutes), year: .y2025)
+            StoryShareableText(L10n.eoyStoryListenedToShareText(ListeningTime2024Story.formattedDayHours(time: listeningTime)), year: .y2025)
         ]
     }
 }
@@ -109,8 +109,8 @@ final private class LottieTextProvider: AnimationKeypathTextProvider {
         self.endTime = endTime
         self.currentTime = startTime
         self.formatter = NumberFormatter()
+        formatter.locale = Locale.current
         formatter.numberStyle = .decimal
-        formatter.groupingSeparator = ","
     }
 
     func text(for keypath: Lottie.AnimationKeypath, sourceText: String) -> String? {