+++ podcasts/End of Year/StoriesConfiguration.swift
@@ -8,6 +8,12 @@ class StoriesConfiguration {
     /// Default value is `false`
     var startOverFromBeginningAfterFinished: Bool = false
 
+    // If set to `true` it will close the playback after the last one finished
+    /// Otherwise, it will just pause on the last one.
+    ///
+    /// Default value is `false`
+    var closeAndDismissAfterFinished: Bool = false
+
     /// The number of stories to preload
     ///
     /// When showing the story number zero, StoriesView will
@@ -21,4 +27,8 @@ class StoriesConfiguration {
     var indicatorHeight: CGFloat = 2
 
     var indicatorSpacing: CGFloat = 2
+
+    var loadingIsTheFirstStory: Bool = false
+
+    var defaultStoriesCount: Int = 7
 }