+++ podcasts/End of Year/StoriesModel.swift
@@ -26,7 +26,7 @@ class StoriesModel: ObservableObject {
             }
 
             // Present share alert if the story is shareable
-            guard dataSource.shareableStory(for: currentStoryIndex) != nil else {
+            guard dataSource.numberOfStories > 0, dataSource.shareableStory(for: currentStoryIndex) != nil else {
                 screenshotTaken = false
                 return
             }
@@ -41,7 +41,7 @@ class StoriesModel: ObservableObject {
 
     private let dataSource: StoriesDataSource
     private let publisher: Timer.TimerPublisher
-    private let configuration: StoriesConfiguration
+    let configuration: StoriesConfiguration
     private let progressModel: StoriesProgressModel
 
     private var cancellables = Set<AnyCancellable>()
@@ -51,6 +51,8 @@ class StoriesModel: ObservableObject {
         currentStory?.duration ?? 0
     }
 
+    private var loadingCancellable: Cancellable?
+
     private var currentStoryIdentifier: String = ""
 
     private var currentStoryIsPlus = false
@@ -58,8 +60,6 @@ class StoriesModel: ObservableObject {
     private var manuallyChanged = false
 
     private var currentStory: Story?
-
-    private var pendingPlaybackShareEvents: [[String: String]] = []
 
     var numberOfStories: Int {
         dataSource.numberOfStories
@@ -99,10 +99,7 @@ class StoriesModel: ObservableObject {
             self?.shareAlertVisibilityChanged(isPresented)
         }
 
-        Task.init {
-            await isReady = dataSource.isReady()
-            failed = !isReady
-        }
+        refresh()
 
         subscribeToNotifications()
     }
@@ -111,9 +108,33 @@ class StoriesModel: ObservableObject {
         isReady = false
 
         Task.init {
-            await isReady = dataSource.refresh()
+            if self.configuration.loadingIsTheFirstStory {
+                loadingStart()
+            }
+            let isReady = await dataSource.refresh()
+            if configuration.loadingIsTheFirstStory, progress >= 1 {
+                currentStoryIndex = 1
+                Analytics.track(.endOfYearStoryShown, story: "cover")
+            }
+            self.isReady = isReady
             failed = !isReady
         }
+    }
+
+    func loadingStart() {
+        loadingCancellable = publisher.autoconnect().sink(receiveValue: { _ in
+            let newProgress = self.progress + (0.01 / EndOfYear.defaultDuration)
+            if newProgress < 1 {
+                self.progress = newProgress
+            } else {
+                self.progress = 1
+            }
+        })
+    }
+
+    func loadingEnded() {
+        loadingCancellable = nil
+        self.progress = 1.01
     }
 
     func start() {
@@ -130,6 +151,10 @@ class StoriesModel: ObservableObject {
                 if self.configuration.startOverFromBeginningAfterFinished {
                     newProgress = 0
                     self.currentStoryIndex = 0
+                }
+                else if self.configuration.closeAndDismissAfterFinished {
+                    Analytics.track(.endOfYearStoriesDismissed, properties: ["source": "auto_progress"])
+                    self.stopAndDismiss()
                 }
                 else {
                     self.pause()
@@ -157,6 +182,9 @@ class StoriesModel: ObservableObject {
         // Otherwise, the paywall appears in front of the story
         if currentStory?.identifier != story.identifier, !story.plusOnly || isPaidUser() {
             story.onAppear()
+            if story.shouldPause {
+                pause()
+            }
         }
 
         currentStory = story
@@ -285,7 +313,6 @@ class StoriesModel: ObservableObject {
 
     func stopAndDismiss() {
         pause()
-        trackPendingPlaybackSharesIfNeeded()
         NavigationManager.sharedManager.dismissPresentedViewController()
     }
 
@@ -329,7 +356,7 @@ class StoriesModel: ObservableObject {
     }
 
     func recordPlaybackShare(properties: [String: String]) {
-        pendingPlaybackShareEvents.append(properties)
+        Analytics.track(.playbackShared, properties: properties)
     }
 }
 
@@ -370,21 +397,23 @@ private extension StoriesModel {
                 pause()
                 screenshotTaken = true
 
-                if dataSource.shareableStory(for: currentStoryIndex) != nil {
+                if dataSource.numberOfStories > 0, dataSource.shareableStory(for: currentStoryIndex) != nil {
                     let year = EndOfYear.currentYear.literalValue
                     let story = currentStoryIdentifier
-                    let properties = ["story": story, "year": year, "from": "screenshot"]
+                    let properties = ["story": story, "current_year": year, "from": "screenshot"]
                     Analytics.track(.endOfYearStoryShared, properties: properties)
                 }
             }
             .store(in: &cancellables)
 
-        ServerNotifications.iapPurchaseCompleted.publisher()
-        .receive(on: DispatchQueue.main)
-        .sink { [weak self] _ in
-            self?.refresh()
+        if EndOfYear.currentYear != .y2025 {
+            ServerNotifications.iapPurchaseCompleted.publisher()
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] _ in
+                self?.refresh()
+            }
+            .store(in: &cancellables)
         }
-        .store(in: &cancellables)
     }
 
     func isPaidUser() -> Bool {
@@ -402,16 +431,6 @@ private extension StoriesModel {
     /// Whether some Plus stories should be skipped or not
     func shouldSkipPlusStories() -> Bool {
         !isPaidUser() && !manuallyChanged && currentStoryIsPlus && nextStoryIsPlus()
-    }
-
-    func trackPendingPlaybackSharesIfNeeded() {
-        guard !pendingPlaybackShareEvents.isEmpty else { return }
-
-        pendingPlaybackShareEvents.forEach {
-            Analytics.track(.playbackShared, properties: $0)
-        }
-
-        pendingPlaybackShareEvents.removeAll()
     }
 }